Implement Kikori as a mode
